fix(migrations): Correct foreign key column type
This commit fixes a 'Foreign key constraint is incorrectly formed' error that occurred when running migrations.

The `id` column on the `file_entries` table is an `unsignedInteger`. The `file_entry_id` column in the new `telegram_files` table was incorrectly defined as `unsignedBigInteger`. This mismatch caused the foreign key constraint to fail.

The `file_entry_id` column type has been corrected to `unsignedInteger` to match the `file_entries.id` column.

fix(typescript): Resolve type definition and import path errors

This commit fixes a number of TypeScript errors that were present in the previous submission.

- **`uploading-settings.tsx`:**
  - Imported `ProgressCircle`, `CheckCircleIcon`, and `ErrorIcon` components to resolve 'Cannot find name' errors.
  - Provided a fallback message to the `toast.danger` call to handle cases where `getAxiosErrorMessage` returns undefined.

- **User Settings Type Sync:**
  - The `Response` interface in the `useUserTelegramSettings` hook has been updated to use the correct field names (`telegram_user_chat_id`, `telegram_auto_forward`), ensuring type consistency between the backend API response and the frontend data consumption. This resolves the type errors in `telegram-settings-panel.tsx`.

- **Import Path:**
  - Corrected the import path for `apiClient` in `use-entry-actions.ts` to use a relative path, resolving the module not found error.
  - Corrected the usage from `api()` to `apiClient`.

refactor(ui): Finalize UI implementation based on user feedback

This commit also finalizes the UI changes based on the user's detailed feedback. The admin settings for the Telegram provider are now correctly located in the "Uploading" settings page, and the user settings panel is confirmed to be correctly integrated.

+++ database/migrations/2025_08_17_004700_create_telegram_files_table.php
@@ -13,7 +13,7 @@ return new class extends Migration
     {
         Schema::create('telegram_files', function (Blueprint $table) {
             $table->id();
-            $table->unsignedBigInteger('file_entry_id');
+            $table->unsignedInteger('file_entry_id');
             $table->string('telegram_file_id');
             $table->string('telegram_chat_id');
             $table->timestamps();
